requirements.txt
openai>=1.50.0,<2.0.0
scikit-learn>=1.4.0,<2.0.0
sentence-transformers>=2.3.1,<3.0.0
numpy>=1.26.3,<2.0.0
flask>=3.0.1,<4.0.0
python-dotenv>=1.0.1
tabulate>=0.9.0
flask-limiter>=3.5.0
flask-cors>=4.0.0
